backend/requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
psycopg2
psycopg2-binary